core/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pika==1.3.1
requests==2.32.3
PyJWT==2.9.0
drf-yasg==1.21.7